agent_lab/agent_lab/aws_docs_agent/requirements.txt
requests
boto3
strands-agents 
strands-agents-tools
bedrock-agentcore-starter-toolkit
uv